+++ clj/src/limabean/adapter/synthetic_spans.clj
@@ -1,42 +1,11 @@
 (ns limabean.adapter.synthetic-spans
-  (:require [limabean.core.format :as format]
-            [clojure.string :as str]
-            [limabean.adapter.pod :as pod]))
+  (:require [clojure.string :as str]
+            [limabean.adapter.pod :as pod]
+            [limabean.core.format :as format]
+            [limabean.util :as util]))
 
-(defn- pop-span!
-  "Pop the first span from the front of the volatile spans"
-  [spans]
-  (let [[span & remaining] @spans]
-    (vreset! spans remaining)
-    span))
-
-(defn- merge-with-directives
-  "Stateful transducer to merge synthetic spans back in with directives"
-  [synthetic-spans]
-  (fn [rf]
-    (let [spans (volatile! synthetic-spans)]
-      (fn
-        ;; init
-        ([] (rf))
-        ;; completion
-        ([result] (rf result))
-        ;; step
-        ([result dct]
-         (if (:provenance dct)
-           (let [dct' (if (= (:dct dct) :txn)
-                        ;; merge txn along with its posting spans
-                        (assoc dct
-                          :span (pop-span! spans)
-                          :postings (mapv (fn [pst]
-                                            (assoc pst :span (pop-span! spans)))
-                                      (:postings dct)))
-                        ;; non-transaction
-                        (assoc dct :span (pop-span! spans)))]
-             (rf result dct'))
-           (rf result dct)))))))
-
-(defn create-with-provenance
-  "Create synthetic spans for all directives/postings with provenance, if required."
+(defn- create-with-provenance
+  "Create synthetic spans for all directives/postings with provenance."
   [directives pod]
   (let [synthetic-span-requests
           (into []
@@ -51,10 +20,25 @@
       (pod/create-synthetic-spans pod synthetic-span-requests)
       [])))
 
-(defn create-and-merge-with-provenance
-  "Create and merge synthetic spans for all directives/postings with provenance, if required."
+(defn- n-spans
+  "Determine how many synthetic spans need merging"
+  [dct]
+  (if (:provenance dct) (inc (count (:postings dct))) 0))
+
+(defn merge-with-provenance
+  "Create and merge synthetic spans as :span-p for all directives/postings with provenance."
   [directives pod]
   (let [synthetic-spans (create-with-provenance directives pod)]
     (if (seq synthetic-spans)
-      (into [] (merge-with-directives synthetic-spans) directives)
+      (vec (util/map-n n-spans
+                       (fn [dct spans]
+                         (let [dct-span (first spans)
+                               pst-spans (rest spans)]
+                           (cond-> (assoc dct :span-p dct-span)
+                             (seq pst-spans) (assoc :postings
+                                               (mapv #(assoc %1 :span-p %2)
+                                                 (:postings dct)
+                                                 pst-spans)))))
+                       directives
+                       synthetic-spans))
       directives)))