+++ clj/test-plugins/src/limabean/test/plugins/auto_close.clj
@@ -1,8 +1,8 @@
 (ns limabean.test.plugins.auto-close)
 
-(defn raw-xf
-  "Transducer on raw directives to immediately close any account in the config hash set."
-  [{:keys [config options]}]
+(defn- xf
+  "Transducer on raw/booked directives to immediately close any account in the config hash set according to phase."
+  [{:keys [config phase]}]
   (fn [rf]
     (fn
       ;; init
@@ -11,7 +11,9 @@
       ([result] (rf result))
       ;; step
       ([result dct]
-       (if (and (= (:dct dct) :open) (contains? config (:acc dct)))
+       (if (and (= phase (:phase config))
+                (= (:dct dct) :open)
+                (contains? (:accs config) (:acc dct)))
          (do
            ;; emit the original open
            (rf result dct)
@@ -19,3 +21,14 @@
            (rf result {:date (:date dct), :dct :close, :acc (:acc dct)}))
          ;; otherwise emit the original directive, whatever it was
          (rf result dct))))))
+
+
+(defn raw-xf
+  "Transducer on raw directives to immediately close any account in the config hash set if phase is raw."
+  [args]
+  (xf (assoc args :phase :raw)))
+
+(defn booked-xf
+  "Transducer on booked directives to immediately close any account in the config hash set if phase is booked."
+  [args]
+  (xf (assoc args :phase :booked)))